requirements.txt
json
jinja2
python-dotenv
openai